+++ test/unit/cursor-stream.test.ts
@@ -1,3 +1,0 @@
-import "./cursor-stream-01.test.js";
-import "./cursor-stream-02.test.js";
-import "./cursor-stream-03.test.js";